+++ src/viewer/tui/mod.rs
@@ -7,7 +7,7 @@ use std::convert;
 
 use anyhow::Context as _;
 use cursive::view::{Resizable as _, Scrollable as _};
-use cursive::views::{Dialog, EditView, LinearLayout, PaddedView, Panel, SelectView, TextView};
+use cursive::views::{Dialog, EditView, LinearLayout, OnEventView, PaddedView, Panel, SelectView, TextView};
 use cursive::{event, theme, utils::markup};
 use cursive_markup::MarkupView;
 
@@ -119,9 +119,32 @@ impl<'s> TuiManRenderer<'s> {
     }
 
     fn into_view(self) -> impl cursive::View {
+        use cursive::With as _;
+        use cursive::view::scroll::Scroller as _;
+
         let title = format!("{} {}", self.doc_ty.name(), self.doc_name);
-        let scroll = self.layout.scrollable().full_screen();
-        Panel::new(scroll).title(title)
+        let scroll = self.layout.scrollable();
+        let wrapper = scroll
+            .wrap_with(OnEventView::new)
+            .on_pre_event_inner(event::Key::PageUp, |v, _| {
+                let scroller = v.get_scroller_mut();
+                if scroller.can_scroll_up() {
+                    scroller.scroll_up(
+                        scroller.last_outer_size().y.saturating_sub(1),
+                    );
+                }
+                Some(event::EventResult::Consumed(None))
+            })
+            .on_pre_event_inner(event::Key::PageDown, |v, _| {
+                let scroller = v.get_scroller_mut();
+                if scroller.can_scroll_down() {
+                    scroller.scroll_down(
+                        scroller.last_outer_size().y.saturating_sub(1),
+                    );
+                }
+                Some(event::EventResult::Consumed(None))
+            });
+        Panel::new(wrapper.full_screen()).title(title)
     }
 }